// src/lib/types/diff-session.ts
import type * as vscode from 'vscode';
import { SelectionInfo } from './selection-info';

export interface DiffSessionSide {
	textKey: string;
	selectionInfo: SelectionInfo;
	originalText: string;
	tempUri: vscode.Uri;
	conflictNotified: boolean;
}

export interface DiffSession {
	id: string;
	title: string;
	left: DiffSessionSide;
	right: DiffSessionSide;
}